// src/store/actions/types.ts
export enum HEADER_ACTION_TYPES {
  FETCH = 'FETCH',
  FAMILY_SELECTED = 'FAMILY_SELECTED',
  CATEGORY_SELECTED = 'CATEGORY_SELECTED',
  CATEGORY_DETAILS_FETCHED = 'CATEGORY_DETAILS_FETCHED',
  SHOW_SEARCH = 'SHOW_SEARCH',
  HIDE_SEARCH = 'HIDE_SEARCH',
  SHOW_PROFILE_LOGIN = 'SHOW_PROFILE_LOGIN',
  HIDE_PROFILE_LOGIN = 'HIDE_PROFILE_LOGIN',
}

export enum NEWSLETTER_ACTION_TYPES {
  RESET = 'RESET',
  SUBMIT = 'SUBMIT',
  SUBMITTING = 'SUBMITTING',
  ADDED_SUCCESS = 'ADDED_SUCCESS',
  ADDED_FAILURE = 'ADDED_FAILURE',
}

export enum SEARCH_ACTION_TYPES {
  SEARCH = 'SEARCH',
  SEARCHING = 'SEARCHING',
  SEARCH_SUCCESS = 'SEARCH_SUCCESS',
  SEARCH_FAILURE = 'SEARCH_FAILURE',
  CLEAR_SEARCH = 'CLEAR_SEARCH',
}

export enum SIDEBAR_ACTION_TYPES {
  SHOW_ACCOUNT = 'SHOW_ACCOUNT',
  SHOW_CART = 'SHOW_CART',
  SHOW_LOGIN = 'SHOW_LOGIN',
  SHOW_REGISTER = 'SHOW_REGISTER',
  SHOW_PROFILE = 'SHOW_PROFILE',
}